fix(core): improve thread safety and correctness
- Extend mutex scope to cover Render() call in game loop
- Use thread_local RNG to eliminate shared mutable state
- Guard against zero/negative fadeTime in character updates
- Cast to float before multiplication to prevent int overflow
- Remove ShowWindow CWRA check (returns prior visibility, not error)
- Remove redundant CharacterSet::Initialize() calls from Spawn/Update

+++ MatrixRainCore/DensityController.cpp
@@ -64,7 +64,7 @@ int DensityController::GetTargetStreakCount() const
 
     // Apply percentage: targetStreaks = maxPossible * (percentage / 100)
     // But always show at least MIN_STREAKS even at 0%
-    int target = static_cast<int> ((maxPossible * m_percentage) / 100.0f);
+    int target = static_cast<int> (static_cast<float> (maxPossible) * m_percentage / 100.0f);
     return std::max (target, MIN_STREAKS);
 }
